# swagger/swagger_output.json
{
    "openapi": "3.0.1",
    "info": {
        "title": "oneM2M API",
        "description": "oneM2M API 1.0",
        "version": "1.0"
    },
    "servers": [
        {
            "url": "/",
            "description": "Default",
            "variables": {}
        }
    ],
    "paths": {
        "/api/onem2m/devices": {
            "post": {
                "tags": [
                    "AE"
                ],
                "description": "Application Entity 리소스 등록",
                "parameters": [
                    {
                        "name": "X-M2M-Origin",
                        "in": "header",
                        "description": "리소스의 Origin (헤더)",
                        "example": "STest_AE_01",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/AE_Create"
                            }
                        }
                    }
                },
                "responses": {
                    "default": {
                        "description": "AE 리소스 등록",
                        "content": {
                            "application/json": {}
                        }
                    }
                }
            }
        },
        "/api/onem2m/devices/{ae}": {
            "get": {
                "tags": [
                    "AE"
                ],
                "description": "Application Entity 리소스 조회",
                "parameters": [
                    {
                        "name": "X-M2M-Origin",
                        "in": "header",
                        "description": "리소스의 Origin (헤더)",
                        "example": "STest_AE_01",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "ae",
                        "in": "path",
                        "description": "AE resourceName, AE 리소스의 이름",
                        "example": "Test_AE_01",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "default": {
                        "description": "AE 리소스 조회",
                        "content": {
                            "application/json": {}
                        }
                    }
                }
            },
            "put": {
                "tags": [
                    "AE"
                ],
                "description": "Application Entity 리소스 업데이트",
                "parameters": [
                    {
                        "name": "X-M2M-Origin",
                        "in": "header",
                        "description": "리소스의 Origin (헤더)",
                        "example": "STest_AE_01",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "ae",
                        "in": "path",
                        "description": "AE resourceName, AE 리소스의 이름",
                        "example": "Test_AE_01",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/AE_Update"
                            }
                        }
                    }
                },
                "responses": {
                    "default": {
                        "description": "AE 리소스 업데이트",
                        "content": {
                            "application/json": {}
                        }
                    }
                }
            },
            "delete": {
                "tags": [
                    "AE"
                ],
                "description": "Application Entity 리소스 삭제",
                "parameters": [
                    {
                        "name": "X-M2M-Origin",
                        "in": "header",
                        "description": "리소스의 Origin (헤더)",
                        "example": "STest_AE_01",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "ae",
                        "in": "path",
                        "description": "AE resourceName, AE 리소스의 이름",
                        "example": "Test_AE_01",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "default": {
                        "description": "AE 리소스 삭제",
                        "content": {
                            "application/json": {}
                        }
                    }
                }
            }
        },
...
        "/api/onem2m/devices/list/AEs": {
            "get": {
                "tags": [
                    "Discovery"
                ],
                "description": "해당 CSEBase 내의 모든 Application Entity 리소스 조회",
                "parameters": [
                    {
                        "name": "X-M2M-Origin",
                        "in": "header",
                        "description": "리소스의 Origin (헤더)",
                        "example": "STest_AE_01",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "default": {
                        "description": "모든 AE 리소스 조회",
                        "content": {
                            "application/json": {}
                        }
                    }
                }
            }
        },
        "/api/onem2m/devices/list/CNTs": {
            "get": {
                "tags": [
                    "Discovery"
                ],
                "description": "해당 CSEBase 내의 모든 Contaier 리소스 조회",
                "parameters": [
                    {
                        "name": "X-M2M-Origin",
                        "in": "header",
                        "description": "리소스의 Origin (헤더)",
                        "example": "STest_AE_01",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "default": {
                        "description": "모든 CNT 리소스 조회",
                        "content": {
                            "application/json": {}
                        }
                    }
                }
            }
        },
        "/api/onem2m/devices/list/{ae}/CNTs": {
            "get": {
                "tags": [
                    "Discovery"
                ],
                "description": "특정 Application Entity 리소스 하위의 Container 리소스 조회",
                "parameters": [
                    {
                        "name": "X-M2M-Origin",
                        "in": "header",
                        "description": "리소스의 Origin (헤더)",
                        "example": "STest_AE_01",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "ae",
                        "in": "path",
                        "description": "AE resourceName, AE 리소스의 이름",
                        "example": "Test_AE_01",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "default": {
                        "description": "특정 AE 하위 CNT 리소스 조회",
                        "content": {
                            "application/json": {}
                        }
                    }
                }
            }
        },
        "/api/onem2m/devices/list/CINs": {
            "get": {
                "tags": [
                    "Discovery"
                ],
                "description": "해당 CSEBase 내의 모든 contentInstance 리소스 조회 (20개 제한)",
                "parameters": [
                    {
                        "name": "X-M2M-Origin",
                        "in": "header",
                        "description": "리소스의 Origin (헤더)",
                        "example": "STest_AE_01",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "default": {
                        "description": "모든 CIN 리소스 조회 (20개 제한)",
                        "content": {
                            "application/json": {}
                        }
                    }
                }
            }
        },
...
        "/api/onem2m/devices/firmware/{nod}": {
            "post": {
                "tags": [
                    "Firmware (mgmtObj)"
                ],
                "description": "Node 리소스 하위 Firmware (mgmtObj) 리소스 등록",
                "parameters": [
                    {
                        "name": "X-M2M-Origin",
                        "in": "header",
                        "description": "리소스의 Origin (헤더)",
                        "example": "STest_NODE_01",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "nod",
                        "in": "path",
                        "description": "NOD resourceName, NOD 리소스의 이름",
                        "example": "Test_NODE_01",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/FWR_Create"
                            }
                        }
                    }
                },
                "responses": {
                    "default": {
                        "description": "NOD 하위 FWR (MGO) 리소스 등록",
                        "content": {
                            "application/json": {}
                        }
                    }
                }
            }
        },
        "/api/onem2m/devices/firmware/{nod}/{fwr}": {
            "get": {
                "tags": [
                    "Firmware (mgmtObj)"
                ],
                "description": "Node 리소스 하위 Firmware (mgmtObj) 리소스 조회",
                "parameters": [
                    {
                        "name": "X-M2M-Origin",
                        "in": "header",
                        "description": "리소스의 Origin (헤더)",
                        "example": "STest_NODE_01",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "nod",
                        "in": "path",
                        "description": "NOD resourceName, NOD 리소스의 이름",
                        "example": "Test_NODE_01",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "fwr",
                        "in": "path",
                        "description": "FWR resourceName, FWR 리소스의 이름",
                        "example": "Test_FWR_01",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "default": {
                        "description": "NOD 하위 FWR (MGO) 리소스 조회",
                        "content": {
                            "application/json": {}
                        }
                    }
                }
            },
            "put": {
                "tags": [
                    "Firmware (mgmtObj)"
                ],
                "description": "Node 리소스 하위 Firmware (mgmtObj) 리소스 업데이트",
                "parameters": [
                    {
                        "name": "X-M2M-Origin",
                        "in": "header",
                        "description": "리소스의 Origin (헤더)",
                        "example": "STest_NODE_01",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "nod",
                        "in": "path",
                        "description": "NOD resourceName, NOD 리소스의 이름",
                        "example": "Test_NODE_01",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "fwr",
                        "in": "path",
                        "description": "FWR resourceName, FWR 리소스의 이름",
                        "example": "Test_FWR_01",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/FWR_Update"
                            }
                        }
                    }
                },
                "responses": {
                    "default": {
                        "description": "NOD 하위 FWR (MGO) 리소스 업데이트",
                        "content": {
                            "application/json": {}
                        }
                    }
                }
            },
            "delete": {
                "tags": [
                    "Firmware (mgmtObj)"
                ],
                "description": "Node 리소스 하위 Firmware (mgmtObj) 리소스 삭제",
                "parameters": [
                    {
                        "name": "X-M2M-Origin",
                        "in": "header",
                        "description": "리소스의 Origin (헤더)",
                        "example": "STest_NODE_01",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "nod",
                        "in": "path",
                        "description": "NOD resourceName, NOD 리소스의 이름",
                        "example": "Test_NODE_01",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "fwr",
                        "in": "path",
                        "description": "FWR resourceName, FWR 리소스의 이름",
                        "example": "Test_FWR_01",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "default": {
                        "description": "NOD 하위 FWR (MGO) 리소스 삭제",
                        "content": {
                            "application/json": {}
                        }
                    }
                }
            }
        }
    },
    "components": {
        "schemas": {
            "AE_Create": {
                "type": "object",
                "properties": {
                    "rn": {
                        "type": "string",
                        "description": "resourceName, 리소스의 이름",
                        "example": "Test_AE_01"
                    },
                    "api": {
                        "type": "string",
                        "description": "applicationID, AE의 고유 ID",
                        "example": "test"
                    },
                    "lbl": {
                        "type": "array",
                        "description": "labels, 리소스의 Tag 정보",
                        "example": [
                            "test ae 01"
                        ]
                    },
                    "rr": {
                        "type": "boolean",
                        "example": true,
                        "description": "requestReachability, 외부 메시지 수신 여부 (기본값 True)"
                    }
                },
                "required": [
                    "rn",
                    "api",
                    "rr"
                ]
            },
            "AE_Update": {
                "type": "object",
                "properties": {
                    "lbl": {
                        "type": "array",
                        "description": "labels, 리소스의 Tag 정보",
                        "example": [
                            "test_ae_01"
                        ]
                    },
                    "rr": {
                        "type": "boolean",
                        "example": true,
                        "description": "requestReachability, 외부 메시지 수신 여부 (기본값 True)"
                    }
                }
            },
...
            "FWR_Create": {
                "type": "object",
                "properties": {
                    "rn": {
                        "type": "string",
                        "description": "resourceName, 리소스의 이름",
                        "example": "Test_FWR_01"
                    },
                    "lbl": {
                        "type": "array",
                        "description": "labels, 리소스의 Tag 정보",
                        "example": [
                            "firmware"
                        ]
                    },
                    "mgd": {
                        "type": "integer",
                        "description": "mgmtDefinition, mgmt 기능 식별자",
                        "example": 1001
                    },
                    "vr": {
                        "type": "string",
                        "description": "version, 펌웨어 또는 소프트웨어 버전",
                        "example": "0.0.1"
                    },
                    "fwnnam": {
                        "type": "string",
                        "description": "firmwareName, 펌웨어 이름",
                        "example": "Test_Firmware"
                    },
                    "url": {
                        "type": "string",
                        "description": "URL, 펌웨어 이미지 등 경로 URL",
                        "example": "http://xxx.xxx.xxx.xxx:xxxx/firmware"
                    },
                    "ud": {
                        "type": "boolean",
                        "description": "update, 펌웨어 업데이트 여부",
                        "example": true
                    }
                },
                "required": [
                    "rn",
                    "mgd"
                ]
            },
            "FWR_Update": {
                "type": "object",
                "properties": {
                    "lbl": {
                        "type": "array",
                        "description": "labels, 리소스의 Tag 정보",
                        "example": [
                            "firmware"
                        ]
                    },
                    "vr": {
                        "type": "string",
                        "description": "version, 펌웨어 또는 소프트웨어 버전",
                        "example": "0.0.1"
                    },
                    "fwnnam": {
                        "type": "string",
                        "description": "firmwareName, 펌웨어 이름",
                        "example": "Test_Firmware"
                    },
                    "url": {
                        "type": "string",
                        "description": "URL, 펌웨어 이미지 등 경로 URL",
                        "example": "http://xxx.xxx.xxx.xxx:xxxx/firmware"
                    },
                    "ud": {
                        "type": "boolean",
                        "description": "update, 펌웨어 업데이트 여부",
                        "example": true
                    }
                }
            }
        }
    }
}